change: fixed-allocator integer size to unsigned 64 bit integer

+++ src/fixed_allocator.cpp
@@ -15,16 +15,11 @@
 
 struct Fixed_Allocator {
 
-   //
-   // @NOTE: This allocator tuned for this specific CLI program and this
-   // program isn't going to need more than the MAX_U16 (65536) bytes of memory.
-   //
-
-   u16 capacity = 0;
-   u16 size = 0;
+   uint capacity = 0;
+   uint size = 0;
    void *buffer;
 
-   Fixed_Allocator sub_allocator(const u16 total_size) {
+   Fixed_Allocator sub_allocator(const uint total_size) {
       assert(total_size + size > capacity, "bump_allocator failed to create a sub-allocator because capacity is full");
 
       int idx = capacity - total_size;
@@ -36,7 +31,7 @@ struct Fixed_Allocator {
       };
    }
 
-   void *alloc(const u16 bytes) {
+   void *alloc(const uint bytes) {
       assert((size + bytes) > capacity, "bump-allocator allocation failed because capacity is full");
       void *memory = {0};
       memory = (char *)buffer + size;
@@ -59,7 +54,7 @@ struct Fixed_Allocator {
    }
 };
 
-Fixed_Allocator fixed_allocator(const u16 size) {
+Fixed_Allocator fixed_allocator(const uint size) {
    void *memory = mmap(NULL, size, PROT_READ | PROT_WRITE, MAP_PRIVATE | MAP_ANON, -1, 0);
    assert_err(memory == MAP_FAILED, "mmap failed");
